move stuff out of storage
